+ port one-char y/n prompting when not --force
this is quite a bit more code and low-level manipulation needed than in
shell where it's a simple "read -n1 -p <prompt>".  if we only wanted to
get chars (even just one) until a newline it would have been much
easier, but we wanted to get one single character only, a y or n,
without the user needing to use "enter" on the line.

this took a bit of research but fortunately is already familiar from my
C days, where termios manipulation was common