+ port find_into() from install.sh (now find_candidates())
we don't need to use a global array variable and assign into it by
string name, and we don't need to use find(1) with strings representing
arguments for what to find, so we now have one function that populates
the right list of DirEntrys according to the program's invocation name.
this function is much easier to read and simpler than it was in shell.
probably not faster though to be honest, since that one was a single
'find' command which is done in straight C