/ correctly follow symlinks when copying and testing filetype
python copy() defaults to copying the destination, not the symlink
same for DirEntry.is_file(), it only tests the target, not the link
itself, but is_symlink() is also try, so it returns true for both
is_file() and is_symlink()! how can it be both things at the same time?
because it follows in former case, but not latter...