/ fix bogus symlink installed vs proposed sameness test
fix major bug that meant the skip test was bad, although it happened to
be right most of the time because of a double-error:

first, the condition was inverted: same should be if it's equal, but
there was a 'not' in front, which was wrong (it's the same if it's the
same, not if it's *not* the same)

second, the comparison generated an absolute path relative to cwd, which
was in the source dir.  so it would never compare correctly because the
link target would always be different

the combination of both is what left this unnoticed: because in most
cases, it was 'not' the wrong thing, which happened to be the right
thing.

some kind of test suite would have caught this (TODO)