+ args: status: add a -s/--status, exits 0 for no changes, 1 for changed
mainly our use case is when this is being invoked from ansible, we'll be
able to tell if any action was performed or not.  we don't want this to
be the typical operation because when running on cli we want zero status
to mean "it has the desired state now" whether it changed it or not,
this is more in keeping with cli usage, whereas ansible will want to
report if changed and the status flag will be more useful there.

still exists with 70 on error (ie, os.EX_SOFTWARE)